SRC/requirements.txt
asyncpg
discord
aiohttp
pillow